Update package dependencies and replace app icons
This commit updates the "@testing-library/react" to version 16.0.0 and adds a new devDependency "@babel/plugin-proposal-private-property-in-object". It also replaces the favicon and logo files with SVG equivalents and updates relevant references in HTML and manifest files. Changes in "package-lock.json" mirror the dependency changes. The metadata in 'index.html' and 'manifest.json' was also updated to fit the new theme.