Add SubmitButton component and update tests
A new SubmitButton component has been added and used in the welcome message and payment information form section. Refactored form layout in the payment information form for better grouping of form inputs, and added respective styles. The test for FormInput has also been updated to include a case for when isMissing flag is false.